Add `getAvatarFromUser` and `getAvatarFromMessage`
We duplicate the same trivaial code for extracting an avatar's url
either from a User object or a Message object.

This code extracts that into two utility functions and replaces the
duplicate code with function calls. DRY.
