+++ src/common/OwnAvatar.js
@@ -5,8 +5,7 @@ import { connect } from 'react-redux';
 import type { GlobalState, User } from '../types';
 import { getCurrentRealm, getSelfUserDetail } from '../selectors';
 import ImageAvatar from './ImageAvatar';
-import { getFullUrl } from '../utils/url';
-import { getGravatarFromEmail } from '../utils/avatar';
+import { getAvatarFromUser } from '../utils/avatar';
 
 type Props = {|
   user: User,
@@ -24,11 +23,7 @@ type Props = {|
 class OwnAvatar extends PureComponent<Props> {
   render() {
     const { user, size, realm } = this.props;
-    const fullAvatarUrl =
-      typeof user.avatar_url === 'string'
-        ? getFullUrl(user.avatar_url, realm)
-        : getGravatarFromEmail(user.email);
-
+    const fullAvatarUrl = getAvatarFromUser(user, realm);
     return <ImageAvatar avatarUrl={fullAvatarUrl} size={size} shape="rounded" />;
   }
 }
