+++ src/utils/__tests__/avatar-test.js
@@ -1,4 +1,4 @@
-/* @flow strict */
+/* @flow strict-local */
 import { getMediumAvatar, getGravatarFromEmail } from '../avatar';
 
 // avatarUrl can be converted to retrieve medium sized avatars(mediumAvatarUrl) if and only if
