+++ src/webview/html/messageTypingAsHtml.js
@@ -1,19 +1,14 @@
 /* @flow strict-local */
 import template from './template';
 import type { User } from '../../types';
-import { getFullUrl } from '../../utils/url';
-import { getGravatarFromEmail } from '../../utils/avatar';
+import { getAvatarFromUser } from '../../utils/avatar';
 
 const typingAvatar = (realm: string, user: User): string => template`
 <div class="avatar">
   <img
     class="avatar-img"
     data-email="${user.email}"
-    src="${
-      typeof user.avatar_url === 'string'
-        ? getFullUrl(user.avatar_url, realm)
-        : getGravatarFromEmail(user.email)
-    }">
+    src="${getAvatarFromUser(user, realm)}">
 </div>
 `;
